backend/requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.22.0
python-dotenv==1.0.0
sqlmodel==0.0.14
websockets==11.0.3
pytest==8.3.3
python-json-logger==2.0.7
numpy>=1.26.0
xgboost>=2.0.0
torch>=2.2.0
scikit-learn>=1.3.0
tensorboard>=2.12.0
optuna>=3.5.0

# Scheduler for monthly training
APScheduler>=3.10.4

